==> scripts_and_markdowns/4_functional_alpha_diversity_modelling.qmd <==
---
title: "Alpha_diversity_modelling"
author: "Patrick Alexander Walkden"
format: html
editor: visual
---

## Functional Alpha diversity modelling

Now let's get to the modelling. The functional intactness index is obtained as the product of a measure of alpha diversity and functional similarity. Here we take functional richness as a function of the PREDICTS land use and intensity classes and the other associated pressures including human population density, road density ( at 10km), time since 30% converted to human modified land uses, percentage of the landscape that is natural habitat using generalized linear mixed effects models (GLMMs).

```{r}

rm(list = ls())

require(tidyverse) ## data manipulation
require(car) ## post-hoc model analyses
library(boot)
require(lme4) # modelling generalised linear mixed effects models
require(lmerTest) ## model scrutiny
require(ggResidpanel) ## model scrutiny 
require(sjPlot) ## visulise interactions
require(sjmisc) ## ditto
require(ggnewscale)

```

## Load in data

Load in data what we will need is just the alpha diversity dataframe that we collated together in the previous markdown with all the pressures and relevant information that we need to get to modelling.

Hvengaard has two studies within the data so I filter out for one of the studies and keep the one that surveys the bird community more completely.

```{r}

alpha_data <- readRDS("../outputs/alpha_diversity_dataframe.rds") %>% dplyr::filter(SS != "GN1_2010__Hvenegaard 2")

table(alpha_data$LUI)

length(unique(alpha_data$SS))
```

We can see the distribution of the land use and intensity classes of the available sites, it looks like here that Cropland intense use has too few sites to make a reliable estimation of it's impact on functional richness therefore I will combine them together with Cropland Light use to make a category of Cropland High use. I also combine urban sites together because ... Andy had a good reason

```{r}

### merge together some land use classes and filter out those sites that we cannot decide what land use intensity it falls under
alpha_data <- alpha_data %>% dplyr::mutate(LUI = ifelse(grepl(LUI, pattern = "Cropland_Intense use")|
                                                           grepl(LUI, pattern = "Cropland_Light use"), "Cropland_High use", LUI)
                                           # LUI = ifelse(grepl(LUI, pattern = "Pasture_Intense use")|
                                           #                 grepl(LUI, pattern = "Pasture_Light use"), "Pasture_High use", LUI),
                                           # LUI = ifelse(grepl(LUI, pattern = "Urban"), "Urban", LUI)
                                           ) %>%
  dplyr::filter(!grepl(LUI , pattern = "Cannot decide"))


## make LUI a fcator for modelling

alpha_data$LUI <- factor(alpha_data$LUI)

## get the levels in the correct order

land_use_levels <- c("Primary forest_Minimal use","Primary forest_Light use","Primary forest_Intense use",
                     "Primary non-forest_Minimal use","Primary non-forest_Light use", "Primary non-forest_Intense use",
                     "Secondary vegetation_Minimal use","Secondary vegetation_Light use","Secondary vegetation_Intense use",
                     "Plantation forest_Minimal use","Plantation forest_Light use","Plantation forest_Intense use",
                     "Pasture_Minimal use","Pasture_Light use","Pasture_Intense use",
                     "Cropland_Minimal use", "Cropland_High use",
                     "Urban_Minimal use", "Urban_Light use","Urban_Intense use")

## relvel LUI factor
alpha_data$LUI <- factor(alpha_data$LUI, levels = land_use_levels)



table(alpha_data$LUI)

```

## Transformations, colinearity and scaling

Before modelling it is good practice to assess the distribution of both the predictor and response variables and to transform them if necessary. Fortunately, the predictor variables have been transformed pre-extraction so it just leaves us to have a look at functional richness.

However, with the predictor variables we will have to check whether there is any colinearity because if there is each variable will be trying to explain the same thing and the variation each explains may be split leading to some unreliable and misleading results.

```{r}

## start with functional richness
hist(alpha_data$functional_richness)

## yeha that needs a transformation

hist(log(alpha_data$functional_richness))
hist(sqrt(alpha_data$functional_richness))

## logging it looks much better as compared to square rooting

alpha_data$logRich <- log(alpha_data$functional_richness) 




## load in package that has functions that has a look at the variance inflation factors of predictor variables - a score less than 10, but ideally 3 means there is little colinearity and modelling is good to continue.

source("https://highstat.com/Books/Book2/HighstatLibV10.R")
corvif(alpha_data[,c( "LUI", "log_hpd_1km", "log_roads","nat_hab_sw" ,"log_T30")])

## brilliant our variables are not colinear.
```

Brilliant there is no colinearity in the predictor variables, and finally before modelling it is good to scale the predictor variables to a mean of zero and a sd of 2 (instead of the usual 1), this enables better interpretation of the results.

```{r}


### FUNCTION

## INPUT: dataframe column that you want to scale ; the number of sds you want to scale to

## OUTPUT: dataframe column with scaled values adn attributes needed to unscale variables that is going to be necessary when projecting the model.

new_scale <- function(x, sds){
  
  mean <- mean(x)
  sd <- sd(x)
  
  
  scale_variable <- (x - mean)/(sds*sd)
  
  
  attr(scale_variable, "scale.scale") <- sds*sd
  attr(scale_variable, "scale.center") <- mean
  
  
  return(scale_variable)
  
}



alpha_data$log_hpd_1km <- new_scale(alpha_data$log_hpd_1km, sds = 2)
alpha_data$log_roads <- new_scale(alpha_data$log_roads, sds = 2)
alpha_data$nat_hab_sw <- new_scale(alpha_data$nat_hab_sw, sds = 2)
alpha_data$log_T30 <- new_scale(alpha_data$log_T30, sds = 2)
alpha_data$control_hpd <- new_scale(alpha_data$control_hpd, sds = 2)
alpha_data$control_roads <- new_scale(alpha_data$control_roads, sds = 2)

```

# Modelling

So now all the variables as transformed and scaled to our satisfaction we are now ready to begin modelling. We are going to start with the maximal model with all main effects and their interactions with land use and intensity to begin with. We will then proceed through backwards step-wise model selection, dropping interactions that are not significant and then main effects. If a main effect has a significant interaction it cannot then be dropped as a main effect.

We are also using generalised mixed effect modeling that enables us to take study as a random effect to account for the significant difference in each studies' region, sampling method sampling effort and other broad effects.

```{r}

## start with the full model with all varaibles included and all interactions
model_1 <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads + LUI:log_T30 +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 |SS/SSB) ,
    data = alpha_data
  )
summary(model_1)
car::Anova(model_1)
BIC(model_1) ## 4054.962
### check for random slopes


model_1a <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads + LUI:log_T30 +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 + log_hpd_1km|SS/SSB) ,
    data = alpha_data
  )
summary(model_1a)
car::Anova(model_1a)
BIC(model_1a) ## 4075.31 NOPE
## roads
model_1b <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads + LUI:log_T30 +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 + log_roads |SS/SSB) ,
    data = alpha_data
  )
summary(model_1b)
car::Anova(model_1b)
BIC(model_1b) ## 4075.852 NOPE
## nat_hab
model_1c <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads + LUI:log_T30 +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 + nat_hab_sw |SS/SSB) ,
    data = alpha_data
  )
summary(model_1c)
car::Anova(model_1c)
BIC(model_1c) ## 4071.29 NOPE
## log_t30
model_1d <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads + LUI:log_T30 +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 + log_T30 |SS/SSB) ,
    data = alpha_data
  )
summary(model_1d)
car::Anova(model_1d)
BIC(model_1d) ## 4074.969 NOPE
### right proceed with the random effect structure of just the random intercepts of study without random slopes for any of the ocntinuous variables
summary(model_1)
car::Anova(model_1)
### interactions between LUI and log roads and log_T30 are very close to being signifant but are not so we'll drop LUI:log_T30 first and see the impact
model_2 <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads +
      log_hpd_1km + log_roads + nat_hab_sw + log_T30 + control_hpd + control_roads +
      (1 |SS/SSB) ,
    data = alpha_data
  )
summary(model_2)
BIC(model_2) ## 3896.766 BIC is a lot lower in model 2 so good to proceed
anova(model_2, model_1)
### the LRT of the models is almost significant meaning that we dont lose a significnat amount of explanatory power (just) so it is good to proceed with the simpler model
car::Anova(model_2)
### This means that we should get rid of log_T30 main effect from the model as it is the lest signifcant variable in the model and is not contained within an interactions.
model_3 <-
  lmer(
    logRich ~ LUI  + LUI:log_hpd_1km + LUI:nat_hab_sw + LUI:log_roads +
      log_hpd_1km + log_roads + nat_hab_sw + control_hpd + control_roads +
      (1 | SS/SSB) ,
    data = alpha_data
  )
summary(model_3)
BIC(model_3) ## 3883.608 is lower by 3 which is not a lot but let's see how
anova(model_3, model_2) ## there is not a signifcnat different in the models explantory power therefore we are good to proceed with the simpler model
car::Anova(model_3)
### looks like we can also drop the interaction between LUI and log roads

```

\

```{r}
ggResidpanel::resid_panel(model_3)

```

ah the diagnostic plots look reasonably okay... the QQ plot looks a bit skew-wif in the lower values but I think this is as good as we're likely to get so I'm happy to go forward.

So that's our final model decided and it looks pretty much as we'd expect. In the absence of the impact of intercations pretty much all other land uses exhibit a decrease in functional richness (apart from cropland minimal and urban - maybe has big interactions with other variables)

```{r}
plot_model(model_3, type = "int", terms = c("LUI", "nat_hab_sw"))



post_hoc_pairwise_comparisons <-   emmeans::emmeans(model_3, pairwise ~ LUI, adjust = "tukey")

contrasts <- summary(post_hoc_pairwise_comparisons)[["contrasts"]]

```

```{r}
## lets save the model and data ready to be projected 

write_rds(file = "../outputs/alpha_diversity_rao_model.rds", x = model_3)
write_rds(file = "../outputs/alpha_modelling_dataframe.rds", x = alpha_data)

```

## Bootstrapping the models

```{r}




scale_object <- function(values, by_column) {
  x <- values - attr(by_column, "scale.center")
  y <- x / attr(by_column, "scale.scale")
  
  return(y)
  
}




data <-
  data.frame(
    LUI = land_use_levels,
    log_hpd_1km = mean(alpha_data$log_hpd_1km),
    nat_hab_sw = mean(alpha_data$nat_hab_sw),
    log_roads = mean(alpha_data$log_roads),
    control_roads = mean(alpha_data$control_roads),
    control_hpd = mean(alpha_data$control_hpd)
  )

# data <-
#   data.frame(
#     LUI = land_use_levels,
#     log_hpd_1km = scale_object(0,alpha_data$log_hpd_1km),
#     nat_hab_sw = scale_object(1,alpha_data$nat_hab_sw),
#     log_roads = scale_object(0,alpha_data$log_roads),
#     control_roads = scale_object(0,alpha_data$control_roads),
#     control_hpd = scale_object(0,alpha_data$control_hpd)
#   )



land_use_colours <-
  data.frame(
    colours = c(
      "green4",
      "chocolate",
      "olivedrab2",
      "springgreen2",
      "#EBF787",
      "#E3D438",
      "#718879"
    )
  )
rownames(land_use_colours) <- c(
  "Primary forest",
  "Primay non-forest",
  "Secondary vegetation",
  "Plantation forest",
  "Pasture",
  "Cropland",
  "Urban"
)




pred_fun <- function(x) {
  as.numeric(predict(x, newdata = data, re.form = NA))
}








data$estimate <- NA
data$upper <- NA
data$lower <- NA


booted_mod <-
  bootMer(
    model_3,
    FUN = function(x)
      pred_fun(x),
    nsim =  200,
    seed = 1234
  )



data$estimate <- exp(booted_mod$t0)
bootstraps <- exp(booted_mod$t)


confidence_intervals <- function(data, highlow) {
  if (highlow == "high") {
    return(data[order(data)[195]])
    
  } else {
    return(data[order(data)[5]])
  }
}


data$upper <-
  apply(
    X = bootstraps,
    MARGIN = 2,
    FUN = function(x)
      confidence_intervals(data = x, highlow = "high")
  )
data$lower <-
  apply(
    X = bootstraps,
    MARGIN = 2,
    FUN = function(x)
      confidence_intervals(data = x, highlow = "low")
  )



data$land_use <-
  unlist(strsplit(data$LUI, split = "_"))[seq(1, by = 2, length.out = nrow(data))]
data$intensity <-
  unlist(strsplit(data$LUI, split = "_"))[seq(2, by = 2, length.out = nrow(data))]

data$LUI <- factor(data$LUI, levels = land_use_levels)


data$land_use <- factor(
  data$land_use,
  levels = c(
    "Primary forest",
    "Primary non-forest",
    "Secondary vegetation",
    "Plantation forest",
    "Pasture",
    "Cropland",
    "Urban"
  )
)






data$intensity <-
  factor(data$intensity,
         levels = c("Minimal use", "Light use", "High use", "Intense use"))


background <- data.frame(
  ystart = seq(0.5, by = 1, length.out = 7),
  yend = seq(1.5, by = 1, length.out = 7),
  col = factor(levels(data$land_use))
)




land_use_effect <-
  ggplot() +
  geom_hline(yintercept = data[data$LUI == "Primary forest_Minimal use", "estimate"], linetype = "dashed") +
  geom_errorbar(
    data = data,
    aes(
      ymax = upper,
      ymin = lower,
      x = land_use,
      group = intensity
    ),
    position = position_dodge(0.75),
    width = 0.3,
    show.legend = FALSE,
    linewidth = 2
  ) +
  geom_point(
    data = data,
    aes(
      x = land_use,
      y = estimate,
      fill = land_use,
      colour = land_use,
      group = intensity,
      shape = intensity
    ),
    position = position_dodge(0.75),
    size = 10,
    show.legend = FALSE
  ) +
  scale_shape_manual(name = "intensity", values = c(21, 22, 23, 24)) +
  scale_fill_manual(values = land_use_colours$colours) +
  scale_colour_manual(values = land_use_colours$colours) +
  ggnewscale::new_scale_fill() +
  geom_rect(
    data = background,
    aes(
      xmin = ystart,
      xmax = yend,
      ymin = 0,
      ymax = 3.6,
      fill = col
    ),
    alpha = 0.2,
    show.legend = FALSE
  ) +
  scale_fill_manual(values = c("white", "grey", "white", "grey", "white", "grey", "grey")) +
  scale_y_continuous(breaks = c(0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5),
                     limits = c(0,3.6)) +
  #scale_y_continuous(limits = c(0.2, 2.2),
  #                  breaks = c(0.5,1,1.5,2)) +
  theme(
    panel.background = element_rect(fill = 'white', color = 'white'),
    panel.grid.major = element_line(color = 'white'),
    panel.grid.minor = element_line(color = 'white'),
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    #axis.ticks.y = element_blank(),
    axis.title.x = element_blank(),
    axis.line = element_blank(),
    axis.ticks.x = element_blank(),
    axis.text.x = element_blank()
  )



plot(land_use_effect)


ggsave(
  filename = "../outputs/FII_maps_and_graphs/alpha_land_use_impact.png",
  land_use_effect,
  device = "png",
  height = 5,
  width = 10,
  dpi = 600
)


```

## Variable significance

```{r}


model_output <- coef(summary(model_3))

confidence_intervals <- confint(model_3, method = "boot", nsim = 200, boot.type = "perc", level = 0.95)



confidence_intervals <- cbind(model_output,confidence_intervals[4:85,])


write.csv(file = "../outputs/forest_alpha_model_output.csv", x = confidence_intervals)


non_forest_alpha_data <- alpha_data
non_forest_alpha_data$LUI <- factor(non_forest_alpha_data$LUI,levels = land_use_levels[c(4:6,1:3,7:20)])

non_for_model_3 <- update(model_3, data = non_forest_alpha_data)

non_for_model_output <- coef(summary(non_for_model_3))

confidence_intervals <- confint(non_for_model_3, method = "boot", nsim = 200, boot.type = "perc", level = 0.95)



confidence_intervals <- cbind(non_for_model_output,confidence_intervals[4:85,])


write.csv(file = "../outputs/non_forest_alpha_model_output.csv", x = confidence_intervals)



data <- alpha_data
land_use <- "Primary non-forest_Minimal use"
variable = "nat_hab_sw"


booted_confidence_intervals <- function(data, land_use, variable){

  pred_fun <- function(x) {
  as.numeric(predict(x, newdata = ndat, re.form = NA))
}

  

land_use_data <- data %>% dplyr::filter(LUI == land_use)  

variable_vals <- seq(min(land_use_data[,variable], na.rm = TRUE),
                     max(land_use_data[,variable], na.rm = TRUE), length.out = 50)


  
ndat <- data.frame(LUI = land_use,
    log_hpd_1km = rep(mean(alpha_data$log_hpd_1km),50),
    nat_hab_sw = rep(mean(alpha_data$nat_hab_sw),50),
    log_roads = rep(mean(alpha_data$log_roads),50),
    control_roads = rep(mean(alpha_data$control_roads),50),
    control_hpd = rep(mean(alpha_data$control_hpd),50))

ndat[,variable] <- variable_vals



bb <- bootMer(model_3, FUN= function(x) pred_fun(x),nsim = 200,seed = 1234)




ndat$booted_estimate <- exp(bb$t0)
booted_CI <- apply(bb$t,MARGIN = 2, FUN = function(x) x[order(x)][c(5,195)])


ndat$LC <- booted_CI[1,]
ndat$UC <- booted_CI[2,]
ndat$booted_estimate <- predict(model_3, newdata=ndat,re.form = NA)


colnames(ndat)[which(colnames(ndat) == variable)] <- "x"
colnames(land_use_data)[which(colnames(land_use_data) == variable)] <- "x"



CI_plot <- ggplot(data = land_use_data, aes(x = x, y = logRich)) +
  geom_point() +
  geom_line(data = ndat, aes(x = x, y = booted_estimate), colour = "orange") + 
  geom_line(data = ndat, aes(x = x, y = LC), linetype = "dashed", colour = "blue") + 
  geom_line(data = ndat, aes(x = x, y = UC), linetype = "dashed", colour = "red") + 
  xlab(variable) +
  
  theme_classic()


return(CI_plot)

}




test <- booted_confidence_intervals(data = alpha_data, land_use = "Primary non-forest_Intense use", variable = "log_hpd_1km")

plot(test)
```
